Install deconvolution script.
The provided deconvolution script seems useful enough to install when
the flowdec package is installed.  This patch installs it as the
"deconvolution" command.

The previous file is left in place and works the same way as before
for the sake of not breaking any current usage or examples.  The
README was updated to point to where the bulk of the code resides.